% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caret.R
\name{createDataPartition}
\alias{createDataPartition}
\title{Data Splitting Functions}
\usage{
createDataPartition(y, times = 1, p = 0.5, list = TRUE, groups = min(5,
  length(y)))
}
\arguments{
\item{y}{data.frame.}

\item{times}{integer.}

\item{p}{numeric.}

\item{list}{logical.}

\item{groups}{numeric.}
}
\description{
Data Splitting Functions
}
\seealso{
\code{nearZeroVar} in the caret package.
}

