% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caret.R
\name{createFolds}
\alias{createFolds}
\title{Create cross validation folds.}
\usage{
createFolds(y, k = 10, list = TRUE, returnTrain = FALSE)
}
\arguments{
\item{y}{vector. A vector of stuff.}

\item{k}{integer. Number of folds.}

\item{list}{logical. Whether or not to return a list.}

\item{returnTrain}{logical. Whether or not to return the train data.}
}
\description{
Create cross validation folds.
}

