% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caret.R
\name{createMultiFolds}
\alias{createMultiFolds}
\title{Create multi folds.}
\usage{
createMultiFolds(y, k = 10, times = 5)
}
\arguments{
\item{y}{vector. A vector of stuff.}

\item{k}{integer. Number of folds.}

\item{times}{integer. Number of times.}
}
\value{
The result of calling \code{\link{createFolds}} on each split.
}
\description{
Create multi folds.
}

