% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{data_summary}
\alias{data_summary}
\title{Summarize Data by Groups}
\usage{
data_summary(data, varname, groupnames)
}
\arguments{
\item{data}{A data frame containing the data to be summarized.}

\item{varname}{A string specifying the name of the variable to summarize.}

\item{groupnames}{A character vector specifying the names of the grouping variables.}
}
\value{
A data frame with the means and standard deviations of the specified variable for each group.
}
\description{
This function computes the mean and standard deviation of a specified variable within groups defined by one or more grouping variables.
}
\examples{
data <- data.frame(
  group = rep(c("A", "B"), each = 5),
  value = c(1:5, 6:10)
)
data_summary(data, varname = "value", groupnames = "group")

}
