% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caret.R
\name{nearZeroVar}
\alias{nearZeroVar}
\title{Identification of near zero variance predictors}
\usage{
nearZeroVar(x, freqCut = 95/5, uniqueCut = 10, saveMetrics = FALSE,
  foreach = FALSE, allowParallel = TRUE)
}
\arguments{
\item{x}{data.frame.}

\item{freqCut}{numeric.}

\item{uniqueCut}{integer.}

\item{saveMetrics}{logical.}

\item{foreach}{logical.}

\item{allowParallel}{logical.}
}
\description{
Identification of near zero variance predictors
}
\seealso{
\code{nearZeroVar} in the caret package.
}

