\name{sure_independence_screening}
\alias{sure_independence_screening}
\title{Sure independence screening helper.
TODO: Document better.}
\usage{
sure_independence_screening(dep_var, indep_var, cutoff = 0.05,
  freq_cutoff = 0.05, link = "logit", depth = 1)
}
\arguments{
  \item{dep_var}{numeric vector}

  \item{indep_var}{vector}

  \item{cutoff}{default is 0.05}

  \item{freq_cutoff}{default is 0.05}

  \item{link}{character vector}

  \item{depth}{do not touch}
}
\description{
Sure independence screening helper. TODO: Document better.
}

