% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variable_summmaries.r
\name{variable_summaries}
\alias{variable_summaries}
\title{Compute variable summaries.}
\usage{
variable_summaries(dataframe)
}
\arguments{
\item{dataframe}{data.frame. A dataset to compute summaries from. The}
}
\value{
a list with "missing_percentages", "means", and "standard_deviations"
   for each numeric variable.
}
\description{
Compute variable summaries.
}

