% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caret.R
\name{createDataPartition}
\alias{createDataPartition}
\title{Data Splitting Functions
A series of test/training partitions are created using
  'createDataPartition' while 'createResample' creates one or more
  bootstrap samples. 'createFolds' splits the data into 'k' groups
  while 'createTimeSlices' creates cross-validation sample
  information to be used with time series data.}
\usage{
createDataPartition(y, times = 1, p = 0.5, list = TRUE, groups = min(5,
  length(y)))
}
\arguments{
\item{y}{data.frame.}

\item{times}{integer.}

\item{p}{numeric.}

\item{list}{logical.}

\item{groups}{numeric.}
}
\description{
Check out documentation for `caret::createFolds`
}
\seealso{
\code{nearZeroVar} in the caret package.
}

