\name{sure_independence_screening}
\alias{sure_independence_screening}
\title{Sure independence screening helper.
TODO: Document better.}
\usage{
sure_independence_screening(dep_var, indep_var, cutoff = 0.05,
  freq_cutoff = 0.05, link = "logit", depth = 1)
}
\description{
Sure independence screening helper. TODO: Document better.
}

